+++ src/tracks/TrackCoverage.java
@@ -287,6 +287,13 @@ public class TrackCoverage extends Track {
 			if(this.isNoFormat()){
 				faSeqStr += base;
 			} else { 
+				faSeqStr += "\033[48;5;" + Config.getColor(ConfigKey.background) + ";38;5;";
+				     if(base == 'A') { faSeqStr += Config.getColor(ConfigKey.seq_a);} 
+				else if(base == 'C') { faSeqStr += Config.getColor(ConfigKey.seq_c);} 
+				else if(base == 'G') { faSeqStr += Config.getColor(ConfigKey.seq_g);} 
+				else if(base == 'T') { faSeqStr += Config.getColor(ConfigKey.seq_t);} 
+				else { faSeqStr += Config.getColor(ConfigKey.seq_other); }
+				faSeqStr += "m" + base + "\033[0m\033[38;5;0;48;5;" + Config.getColor(ConfigKey.background) + "m"; // Clear formatting and fg to black and bg to white;
 			}
 		}
 		if(allEmpty){